moved converter to separate file and sped it up
