+++ api/converters/README.md
@@ -1,0 +1,2 @@
+This directory is for python-based converters, to convert various RF recording formats to SigMF (and maybe even the other way!)
+They are used within app/converter_router.py to expose the converters to the IQEngine website, but can also be used directly
